src/app/dialog/dialog.component.ts
import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';

@Component({
  selector: 'app-dialog',
  templateUrl: './dialog.component.html',
  styleUrls: ['./dialog.component.css']
})
export class DialogComponent implements OnInit {
  @Input() value: string;
  @Input() showPrompt: boolean;
  @Input() placeholder: string;
  @Input() title: string;
  @Input() okText: string;
  @Input() cancelText: string;
  @Input() template: string;

  @Output()  valueEmitted = new EventEmitter<string>();

  constructor() {
    this.okText = 'OK';
    this.cancelText = 'Cancel';
    this.value = null;
   }

  ngOnInit() {
  }

  emitValue(value) {
   this.valueEmitted.emit(value);
  }
}
